Wrote and tested a shell script that adds a file
For example, in this case, I typed "add-file.sh b.txt". So, the file b.txt
was created, it was added to the list of files opened by
open-my-files.bat, and a commit was generated with the message, "Added new
file: b.txt". (However, before calling the script, I accidentally staged
the script, so it was added to that commit, which is why I've amended this
commit. Otherwise, the commit would've been properly generated.)