Added a local version of jQuery
Thus, I'm temporarily not using copy delivered by Google Content Delivery
Network. Yesterday, I was having many problems debugging dealOutCards()
(the implementation to which I will shortly commit), so, perhaps
irrationally, I made the change in this commit to narrow down the source
of possible problems. During the debugging, I'd open the browser to find
it constantly in a loading state, which made it plausible that Google CDN
could be the problem (at least, to me at the time). Later, I'll either
undo this change or try to implement a fallback (in case Google CDN
fails).