Implemented dealOutCards() and its unit tests
Due to a bug(s), it took me many hours to implement this function, yet the
problem wasn't too complicated. The problem was that I hadn't realized
that dealOutCards() will infinitely loop if bs.players is empty. Thus,
I've updated the precondition and artificially fixed this problem in the
document ready function in bs.js.