Play order is randomzied
However, this may not be the best way to determine play order. As I said
on the todo list, it may be more fair to base play order on the order in
which the cards were dealt, so that those who have more cards will go
before those who have less cards (since with five players, for example,
the first two players will each have one more card than each of the other
three players has). Yet, with this method, the order could be randomized
before the cards are dealt.