Fixed a bug in randomizePlayerOrder()
This bug was wrapping each's player's cards in another layer of array.
However, I don't see the need to provide a means to catch this bug in the
unit tests, for the only way it could cause a problem again would occur if
someone intentionally deleted the addition I've made in this commit.