Fixed submitCards()
It had a bug that would cause it to not check the last cards held by a
player if enough cards of his were already removed; this had to do with
the way I traversed the for loop. I made the unit test for this function
more exhaustive (i.e. as exhaustive as it should've been).