Each of each player's displayed cards has the correct card image by it
There are still some things to be done. The textual name of each card is
still near the card. Furthermore, the image of a card submitted by a
player isn't shown. Later, I'll probably need to add to each card an
appropriate id, so that the image of a selected card can be edited.