Picked cards are given a red tint
The article that showed me how to do this is called: "Tinted Images with
Multiple Backgrounds" by Chris Coyier. The technique, which uses
gradients, works on the versions of Chrome, Internet Explorer, Opera, and
Firefox that I have; however (without the appropriate prefix, perhaps),
the gradients don't work on Safari (5 for Windows).