Gradients work on Safari
I figured out how to do this with the help of the webpage "CSS3 Gradients"
on w3schools.com. Also, with the help of this article, in the comments of
bs.css, I specified which browsers the gradients work on. As I stated on
the todo list, I'll figure out what to do about browsers on which the
gradients don't work later.